services/fl_management/requirements.txt
typer
python-dotenv
requests
cysystemd
gitpython
humanize
